+++ src/models/restcollectmethodgetstatetracking.ts
@@ -445,6 +445,18 @@ export type RestAuthenticationNone = {
     | discriminatedUnionTypes.Unknown<"type">
     | undefined;
   __scheduling?: RestAuthenticationNoneScheduling | undefined;
+  /**
+   * Secret value to add to HTTP requests as the 'client secret' parameter. Value is stored encrypted on disk and automatically added to request parameters.
+   */
+  clientSecretParamValue?: string | undefined;
+  /**
+   * Contents of Google Cloud service account credentials (JSON keys) file. To upload a file, click the upload icon in this field's upper right.
+   */
+  serviceAccountCredentials?: string | undefined;
+  /**
+   * Select or create an HMAC Function to use with authentication
+   */
+  hmacFunctionId?: string | undefined;
 };
 
 export const RestCollectMethodOtherCollectMethod = {
@@ -1112,6 +1124,18 @@ export type RestCollectMethodOther = {
     | discriminatedUnionTypes.Unknown<"type">
     | undefined;
   __scheduling?: RestCollectMethodOtherScheduling | undefined;
+  /**
+   * Secret value to add to HTTP requests as the 'client secret' parameter. Value is stored encrypted on disk and automatically added to request parameters.
+   */
+  clientSecretParamValue?: string | undefined;
+  /**
+   * Contents of Google Cloud service account credentials (JSON keys) file. To upload a file, click the upload icon in this field's upper right.
+   */
+  serviceAccountCredentials?: string | undefined;
+  /**
+   * Select or create an HMAC Function to use with authentication
+   */
+  hmacFunctionId?: string | undefined;
 };
 
 export const RestCollectMethodPostWithBodyCollectMethod = {
@@ -1773,6 +1797,18 @@ export type RestCollectMethodPostWithBody = {
     | discriminatedUnionTypes.Unknown<"type">
     | undefined;
   __scheduling?: RestCollectMethodPostWithBodyScheduling | undefined;
+  /**
+   * Secret value to add to HTTP requests as the 'client secret' parameter. Value is stored encrypted on disk and automatically added to request parameters.
+   */
+  clientSecretParamValue?: string | undefined;
+  /**
+   * Contents of Google Cloud service account credentials (JSON keys) file. To upload a file, click the upload icon in this field's upper right.
+   */
+  serviceAccountCredentials?: string | undefined;
+  /**
+   * Select or create an HMAC Function to use with authentication
+   */
+  hmacFunctionId?: string | undefined;
 };
 
 export const RestCollectMethodPostCollectMethod = {
@@ -2432,6 +2468,18 @@ export type RestCollectMethodPost = {
     | discriminatedUnionTypes.Unknown<"type">
     | undefined;
   __scheduling?: RestCollectMethodPostScheduling | undefined;
+  /**
+   * Secret value to add to HTTP requests as the 'client secret' parameter. Value is stored encrypted on disk and automatically added to request parameters.
+   */
+  clientSecretParamValue?: string | undefined;
+  /**
+   * Contents of Google Cloud service account credentials (JSON keys) file. To upload a file, click the upload icon in this field's upper right.
+   */
+  serviceAccountCredentials?: string | undefined;
+  /**
+   * Select or create an HMAC Function to use with authentication
+   */
+  hmacFunctionId?: string | undefined;
 };
 
 export const RestCollectMethodGetCollectMethod = {
@@ -4183,6 +4231,9 @@ export const RestAuthenticationNone$inboundSchema: z.ZodType<
   __scheduling: types.optional(
     z.lazy(() => RestAuthenticationNoneScheduling$inboundSchema),
   ),
+  clientSecretParamValue: types.optional(types.string()),
+  serviceAccountCredentials: types.optional(types.string()),
+  hmacFunctionId: types.optional(types.string()),
 });
 /** @internal */
 export type RestAuthenticationNone$Outbound = {
@@ -4226,6 +4277,9 @@ export type RestAuthenticationNone$Outbound = {
     | RestAuthenticationNoneRestRetryRulesTypeBackoff$Outbound
     | undefined;
   __scheduling?: RestAuthenticationNoneScheduling$Outbound | undefined;
+  clientSecretParamValue?: string | undefined;
+  serviceAccountCredentials?: string | undefined;
+  hmacFunctionId?: string | undefined;
 };
 
 /** @internal */
@@ -4288,6 +4342,9 @@ export const RestAuthenticationNone$outboundSchema: z.ZodType<
   ]).optional(),
   __scheduling: z.lazy(() => RestAuthenticationNoneScheduling$outboundSchema)
     .optional(),
+  clientSecretParamValue: z.string().optional(),
+  serviceAccountCredentials: z.string().optional(),
+  hmacFunctionId: z.string().optional(),
 });
 
 export function restAuthenticationNoneToJSON(
@@ -6217,6 +6274,9 @@ export const RestCollectMethodOther$inboundSchema: z.ZodType<
   __scheduling: types.optional(
     z.lazy(() => RestCollectMethodOtherScheduling$inboundSchema),
   ),
+  clientSecretParamValue: types.optional(types.string()),
+  serviceAccountCredentials: types.optional(types.string()),
+  hmacFunctionId: types.optional(types.string()),
 });
 /** @internal */
 export type RestCollectMethodOther$Outbound = {
@@ -6265,6 +6325,9 @@ export type RestCollectMethodOther$Outbound = {
     | RestCollectMethodOtherRestRetryRulesTypeBackoff$Outbound
     | undefined;
   __scheduling?: RestCollectMethodOtherScheduling$Outbound | undefined;
+  clientSecretParamValue?: string | undefined;
+  serviceAccountCredentials?: string | undefined;
+  hmacFunctionId?: string | undefined;
 };
 
 /** @internal */
@@ -6344,6 +6407,9 @@ export const RestCollectMethodOther$outboundSchema: z.ZodType<
   ]).optional(),
   __scheduling: z.lazy(() => RestCollectMethodOtherScheduling$outboundSchema)
     .optional(),
+  clientSecretParamValue: z.string().optional(),
+  serviceAccountCredentials: z.string().optional(),
+  hmacFunctionId: z.string().optional(),
 });
 
 export function restCollectMethodOtherToJSON(
@@ -8294,6 +8360,9 @@ export const RestCollectMethodPostWithBody$inboundSchema: z.ZodType<
   __scheduling: types.optional(
     z.lazy(() => RestCollectMethodPostWithBodyScheduling$inboundSchema),
   ),
+  clientSecretParamValue: types.optional(types.string()),
+  serviceAccountCredentials: types.optional(types.string()),
+  hmacFunctionId: types.optional(types.string()),
 });
 /** @internal */
 export type RestCollectMethodPostWithBody$Outbound = {
@@ -8338,6 +8407,9 @@ export type RestCollectMethodPostWithBody$Outbound = {
     | RestCollectMethodPostWithBodyRestRetryRulesTypeBackoff$Outbound
     | undefined;
   __scheduling?: RestCollectMethodPostWithBodyScheduling$Outbound | undefined;
+  clientSecretParamValue?: string | undefined;
+  serviceAccountCredentials?: string | undefined;
+  hmacFunctionId?: string | undefined;
 };
 
 /** @internal */
@@ -8420,6 +8492,9 @@ export const RestCollectMethodPostWithBody$outboundSchema: z.ZodType<
   __scheduling: z.lazy(() =>
     RestCollectMethodPostWithBodyScheduling$outboundSchema
   ).optional(),
+  clientSecretParamValue: z.string().optional(),
+  serviceAccountCredentials: z.string().optional(),
+  hmacFunctionId: z.string().optional(),
 });
 
 export function restCollectMethodPostWithBodyToJSON(
@@ -10345,6 +10420,9 @@ export const RestCollectMethodPost$inboundSchema: z.ZodType<
   __scheduling: types.optional(
     z.lazy(() => RestCollectMethodPostScheduling$inboundSchema),
   ),
+  clientSecretParamValue: types.optional(types.string()),
+  serviceAccountCredentials: types.optional(types.string()),
+  hmacFunctionId: types.optional(types.string()),
 });
 /** @internal */
 export type RestCollectMethodPost$Outbound = {
@@ -10391,6 +10469,9 @@ export type RestCollectMethodPost$Outbound = {
     | RestCollectMethodPostRestRetryRulesTypeBackoff$Outbound
     | undefined;
   __scheduling?: RestCollectMethodPostScheduling$Outbound | undefined;
+  clientSecretParamValue?: string | undefined;
+  serviceAccountCredentials?: string | undefined;
+  hmacFunctionId?: string | undefined;
 };
 
 /** @internal */
@@ -10466,6 +10547,9 @@ export const RestCollectMethodPost$outboundSchema: z.ZodType<
   ]).optional(),
   __scheduling: z.lazy(() => RestCollectMethodPostScheduling$outboundSchema)
     .optional(),
+  clientSecretParamValue: z.string().optional(),
+  serviceAccountCredentials: z.string().optional(),
+  hmacFunctionId: z.string().optional(),
 });
 
 export function restCollectMethodPostToJSON(